Update Author field for Image model, use mixin for Author methods
